% Generated by roxygen2 (4.0.2): do not edit by hand
\name{animated_state_unemployment_choropleth}
\alias{animated_state_unemployment_choropleth}
\title{Create an animated choropleth of US State Unemployment Data}
\usage{
animated_state_unemployment_choropleth()
}
\description{
Animation is done with the choroplethr_animate function from the choroplethr package.
}

