% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_county_df}
\alias{build_county_df}
\title{Build the data object ?df_county_unemployment}
\usage{
build_county_df()
}
\description{
This function is included to allow you to verify the integrity of ?df_county_unemployment.
This will scrape the Bureau of Labor Statistics Website to get the data.
}

