% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{build_state_df}
\alias{build_state_df}
\title{Build the data object ?df_state_unemployment}
\usage{
build_state_df()
}
\description{
This function is included to allow you to verify the integrity of ?df_state_unemployment.
This will scrape the Bureau of Labor Statistics Website to get the data.
}

