% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{df_county_unemployment}
\alias{df_county_unemployment}
\title{A data.frame US County Unemployment Data}
\usage{
data(df_county_unemployment)
}
\description{
Contains annualized data from 1990-2013. Data comes from the US Bureau
of Labor Statistics (BLS) Local Area Unemployment Statistics (LAUS) webpage: http://www.bls.gov/lau/.
}
\examples{
data(df_county_unemployment)

head(df_county_unemployment)
boxplot(df_county_unemployment[, -1],
        main="USA County Unemployment Data",
        xlab="Year",
        ylab="Percent Unemployment")
}
\author{
Ari Lamstein
}
\references{
Created via build_county_df() on January 2, 2014.
}
\keyword{data}

