% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{df_state_unemployment}
\alias{df_state_unemployment}
\title{A data.frame containing US State Unemployment Rates}
\usage{
data(df_state_unemployment)
}
\description{
Contains annualized data from 2000-2013. Data comes from the US Bureau
of Labor Statistics (BLS) Local Area Unemployment Statistics (LAUS) webpage: http://www.bls.gov/lau/.
}
\examples{
data(df_state_unemployment)

head(df_state_unemployment)
boxplot(df_state_unemployment[, -1],
        main="USA State Unemployment Data",
        xlab="Year",
        ylab="Percent Unemployment")

\dontrun{
state_unemployment_choropleth(year=2013)
}
}
\author{
Ari Lamstein
}
\references{
Created via build_state_df() on January 4, 2015.
}
\keyword{data}

