% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_county_unemployment_df}
\alias{get_county_unemployment_df}
\title{Get county level unemployment data}
\usage{
get_county_unemployment_df(year)
}
\arguments{
\item{year}{The year of the data you want. Must bet between 1990 and 2013}
}
\description{
Data is scraped from the US Bureau of Labor Statistics (BSL) webpage. This function is
intended for internal use only - please use ?df_county_unemployment instead.
}

