% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_state_unemployment_df}
\alias{get_state_unemployment_df}
\title{Get state unemployment data from the US Bureau of Labor Statistics (BLS) website}
\usage{
get_state_unemployment_df(year = 2013)
}
\arguments{
\item{year}{A year (integer) between 2000 and 2013}
}
\description{
This function is included only for reference. Please use the data.frame
?df_state_unemployment instead.
}

