% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_unemployment_choropleth}
\alias{state_unemployment_choropleth}
\title{Render Choropleth Map of US State Unemployment Rates}
\usage{
state_unemployment_choropleth(year = 2013, num_colors = 7, zoom = NULL)
}
\arguments{
\item{year}{The year of data to use. Must be between 2000 and 2013.}

\item{num_colors}{The number of colors on the map. A value of 1 will use a continuous scale. 
A value in [2, 9] will use that many colors.}

\item{zoom}{An optional vector of states to zoom in on. Elements of this vector 
must exactly match the names of states as they appear in the "region" column of 
?state.regions in the choroplethrMaps package.}
}
\description{
Data comes from ?df_state_unemployment. The choropleth is rendered with the function
?state_choropleth in the choroplethr package.
}

