% Generated by roxygen2 (4.0.2): do not edit by hand
\name{animated_state_unemployment_choropleth}
\alias{animated_state_unemployment_choropleth}
\title{Create an animated choropleth of US State Unemployment Data}
\usage{
animated_state_unemployment_choropleth()
}
\description{
Data comes from ?df_state_unemployment. The choropleth is rendered with the function
?state_choropleth in the choroplethr package. Note that this command will write files to
your local file system - see ?choroplethr_animate in the choroplethr package for details.
}

