% Generated by roxygen2 (4.0.2): do not edit by hand
\name{county_unemployment_choropleth}
\alias{county_unemployment_choropleth}
\title{Render a Choropleth Map of US County Unemployment Rates}
\usage{
county_unemployment_choropleth(year = 2013, buckets = 7, zoom = NULL)
}
\arguments{
\item{year}{The year of data to use. Must be between 1990 and 2013.}

\item{buckets}{The number of equally sized buckets to places the values in.
A value of 1 will use a continuous scale, and a value in [2, 9] will use that many buckets.}

\item{zoom}{An optional vector of states to zoom in on. Elements of this vector
must exactly match the names of states as they appear in the "region" column of
?state.regions in the choroplethrMaps package.}
}
\description{
Data comes from ?df_county_unemployment. The choropleth is rendered with the function
?county_choropleth in the choroplethr package.
}

