% Generated by roxygen2 (4.0.2): do not edit by hand
\name{state_unemployment_choropleth}
\alias{state_unemployment_choropleth}
\title{Render a State Map of US Unemployment Rates}
\usage{
state_unemployment_choropleth(year = 2000, buckets = 7, zoom = NULL)
}
\arguments{
\item{year}{The year of data to use. Must be between 2000 and 2013.}

\item{buckets}{The number of equally sized buckets to places the values in.
A value of 1 will use a continuous scale, and a value in [2, 9] will use that many buckets.}

\item{zoom}{An optional vector of states to zoom in on. Elements of this vector
must exactly match the names of states as they appear in the "region" column of
?state.regions in the choroplethrMaps package.}
}
\description{
Data comes from ?df_state_unemployment. The choropleth is rendered with the function
?state_choropleth in the choroplethr package.
}

